fixing metadata
